Fix implementation of totalBytes() & totalRows() for Log family.

+++ src/Common/FileChecker.cpp
@@ -69,9 +69,9 @@ size_t FileChecker::getFileSize(const String & full_file_path) const
     return it->second;
 }
 
-UInt64 FileChecker::getTotalSize() const
+size_t FileChecker::getTotalSize() const
 {
-    UInt64 total_size = 0;
+    size_t total_size = 0;
     for (auto size : map | boost::adaptors::map_values)
         total_size += size;
     return total_size;
